# chat-app/docs/swagger.yaml
definitions:
  errs.Error:
    properties:
      message:
        type: string
    type: object
  errs.ValidationError:
    properties:
      errors:
        additionalProperties:
          type: string
        type: object
      message:
        type: string
    type: object
  handler.ListResponse-models_RoomsOutPutModel:
    properties:
      data:
        items:
          $ref: '#/definitions/models.RoomsOutPutModel'
        type: array
      from:
        type: integer
      last_page:
        type: integer
      limit:
        type: integer
      page:
        type: integer
      to:
        type: integer
      total:
        type: integer
    type: object
  handler.Response-handler_ListResponse-models_RoomsOutPutModel:
    properties:
      data:
        $ref: '#/definitions/handler.ListResponse-models_RoomsOutPutModel'
      message:
        type: string
      status:
        type: integer
    type: object
  handler.Response-uuid_UUID:
    properties:
      data:
        type: string
      message:
        type: string
      status:
        type: integer
    type: object
  models.CreatePrivateRoomInputModel:
    properties:
      userId:
        type: string
    type: object
  models.CreatePublicRoomInputModel:
    properties:
      name:
        type: string
      users:
        items:
          type: string
        type: array
    type: object
  models.RoomsOutPutModel:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
info:
  contact: {}
paths:
  /room:
    get:
      consumes:
      - application/json
      parameters:
      - description: page size
        in: query
        name: page
        type: string
      - description: length of records to show
        in: query
        name: limit
        type: string
      - description: search for item
        in: query
        name: searchTerm
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Response-handler_ListResponse-models_RoomsOutPutModel'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errs.Error'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/errs.ValidationError'
      security:
      - Bearer: []
      tags:
      - room
  /room/delete/{id}:
    post:
      consumes:
      - application/json
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Response-uuid_UUID'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errs.Error'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/errs.ValidationError'
      security:
      - Bearer: []
      tags:
      - room
  /room/join/{id}:
    post:
      consumes:
      - application/json
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Response-uuid_UUID'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errs.Error'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/errs.ValidationError'
      security:
      - Bearer: []
      tags:
      - room
  /room/left/{id}:
    post:
      consumes:
      - application/json
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Response-uuid_UUID'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errs.Error'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/errs.ValidationError'
      security:
      - Bearer: []
      tags:
      - room
  /room/private:
    post:
      consumes:
      - application/json
      parameters:
      - description: input model
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.CreatePrivateRoomInputModel'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Response-uuid_UUID'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errs.Error'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/errs.ValidationError'
      security:
      - Bearer: []
      tags:
      - room
  /room/public:
    post:
      consumes:
      - application/json
      parameters:
      - description: input model
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.CreatePublicRoomInputModel'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Response-uuid_UUID'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errs.Error'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/errs.ValidationError'
      security:
      - Bearer: []
      tags:
      - room
swagger: "2.0"